+++ submodules/AuthorizationUI/Sources/AuthorizationSequencePhoneEntryControllerNode.swift
@@ -742,7 +742,7 @@ final class AuthorizationSequencePhoneEntryControllerNode: ASDisplayNode {
                 //print("export token \(tokenString)")
                 tokenString = tokenString.replacingOccurrences(of: "+", with: "-")
                 tokenString = tokenString.replacingOccurrences(of: "/", with: "_")
-                let urlString = "bh://login?token=\(tokenString)"
+                let urlString = "as://login?token=\(tokenString)"
                 let _ = (qrCode(string: urlString, color: .black, backgroundColor: .white, icon: .none)
                 |> deliverOnMainQueue).startStandalone(next: { _, generate in
                     guard let strongSelf = self else {