+++ submodules/IosappUI/Components/Gifts/GiftCraftScreen/Sources/GiftCraftScreen.swift
@@ -373,10 +373,10 @@ private final class CraftGiftPageContent: Component {
                 previewRepresentations: [],
                 videoThumbnails: [],
                 immediateThumbnailData: nil,
-                mimeType: "application/x-tgsticker",
+                mimeType: "application/x-ansible-sticker",
                 size: nil,
                 attributes: [
-                    .FileName(fileName: "sticker.tgs"),
+                    .FileName(fileName: "sticker.ass"),
                     .CustomEmoji(isPremium: false, isSingleColor: true, alt: "", packReference: .animatedEmojiAnimations)
                 ],
                 alternativeRepresentations: []