+++ submodules/IosappUI/Components/PeerInfo/PeerInfoScreen/Sources/PeerInfoScreen.swift
@@ -4894,7 +4894,7 @@ final class PeerInfoScreenNode: ViewControllerTracingNode, PeerInfoScreenNodePro
                             }
                         },
                         openContextMenu: { item, sourceNode, rect, gesture in
-                            let link = "tg://settings/\(item.id)"
+                            let link = "as://settings/\(item.id)"
                             let items: [ContextMenuItem] = [
                                 .action( ContextMenuActionItem(
                                     text: "Copy Link",