+++ submodules/OpenInExternalAppUI/Sources/OpenInOptions.swift
@@ -230,12 +230,12 @@ private func allOpenInOptions(context: AccountContext, item: OpenInItem) -> [Ope
                         case .transit:
                             directionsMode = "transit"
                     }
-                    return .openUrl(url: "comgooglemaps-x-callback://?daddr=\(coordinates)&directionsmode=\(directionsMode)&x-success=telegram://?resume=true&x-source=Telegram")
+                    return .openUrl(url: "comgooglemaps-x-callback://?daddr=\(coordinates)&directionsmode=\(directionsMode)&x-success=ansible://?resume=true&x-source=Ansible")
                 } else {
                     if let venue = location.venue, let venueId = venue.id, let provider = venue.provider, provider == "gplaces" {
                         return .openUrl(url: "https://www.google.com/maps/search/?api=1&query=\(venue.address ?? "")&query_place_id=\(venueId)")
                     } else {
-                        return .openUrl(url: "comgooglemaps-x-callback://?center=\(coordinates)&q=\(coordinates)&x-success=telegram://?resume=true&x-source=Telegram")
+                        return .openUrl(url: "comgooglemaps-x-callback://?center=\(coordinates)&q=\(coordinates)&x-success=ansible://?resume=true&x-source=Ansible")
                     }
                 }
             }))