+++ submodules/UrlWhitelist/Sources/UrlWhitelist.swift
@@ -84,7 +84,7 @@ public func parseUrl(url: String, wasConcealed: Bool) -> (string: String, concea
         concealed = false
     }
     
-    if url.hasPrefix("tg://premium_multigift") || url.hasPrefix("tg://premium_offer") {
+    if url.hasPrefix("as://premium_multigift") || url.hasPrefix("as://premium_offer") {
         concealed = false
     }